Extract core async dispatcher from ActiveJobDispatcher
